Change ConfusionMatrix accumulator from Float to Long.

+++ ConfusionMatrix.lua
@@ -17,7 +17,7 @@ function ConfusionMatrix:__init(nclasses, classes)
       classes = nclasses
       nclasses = #classes
    end
-   self.mat = torch.FloatTensor(nclasses,nclasses):zero()
+   self.mat = torch.LongTensor(nclasses,nclasses):zero()
    self.valids = torch.FloatTensor(nclasses):zero()
    self.unionvalids = torch.FloatTensor(nclasses):zero()
    self.nclasses = nclasses
